Added Placeholder Menu and made a lighting circl
Placeholder menu added, need to replace with assets and do some touching up.

Want the obj_Player to be "falling" but staying dead center of screen.

I made a lighting circle for the player object but couldnt get the rest of of the levels to black out. Will continue to work on this so that the player object is the light source.